Add USDT to the list of fiat currencies
USDT is common many exchanges and treated as a currency. 
It is helpful to allow USDT to be accepted as a fiat current when testing config.json
 
By way of example of a usecase
A strategy or otherwise report may not query:  config['stake_currency'] and currency = config['fiat_display_currency']
as the config.json with fail test, so cannot build  a base/pair such as BTC/USDT that is valid will not be built cleanly. 

Pulling BTC/USDT from config values can be useful to download query ticker file name and data, enumerating base/pair price to whitelist target pairs, performing correlation coefficient statistics etc.

As example: building a dataframe such as here 
           date       BTC/USDT      ETH/BTC          XLM/BTC        LTC/BTC
0    1517443200       10283.00       0.109475       0.000052       0.016031

The transposing to actual USDT ticker value row-wise per pair
               date  BTC/USDT   ETH/USD   XLM/USD     LTC/USD
1517443200  10283.00   1125.731425  0.535950  164.846773