prevent index out of range in trim_tickerlist
There is a bug in trim_tickerlist, it will throw an index out of range exception and crash the programe if passed a start / stop date outside the range of those records in the cached json pair file.

This change calls a function that sanitizes the dates prior to calling trim_tickerlist.