Conception problem with the unrealized profit
I think here is a conception problem with the total profit, if we *= it with the last PnL it means we assume that the last profit effects on this one. In real life it depends on the user eg stake amount can be unlimited.
However the agent will learn better and more if we just add it with +=.